add quads compliance tests
