--- qendpoint-store/src/test/java/com/the_qa_company/qendpoint/store/EndpointSPARQL11UpdateComplianceTest.java
@@ -1,63 +1,0 @@
-package com.the_qa_company.qendpoint.store;
-
-import com.the_qa_company.qendpoint.core.hdt.HDT;
-import com.the_qa_company.qendpoint.core.options.HDTOptions;
-import com.the_qa_company.qendpoint.core.options.HDTOptionsKeys;
-import org.eclipse.rdf4j.model.IRI;
-import org.eclipse.rdf4j.repository.Repository;
-import org.eclipse.rdf4j.repository.sail.SailRepository;
-import org.eclipse.rdf4j.testsuite.query.parser.sparql.manifest.SPARQL11UpdateComplianceTest;
-import org.junit.Rule;
-import org.junit.rules.TemporaryFolder;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-/**
- * Test SPARQL 1.1 Update functionality on a native store.
- *
- * @author Ali Haidar
- */
-public class EndpointSPARQL11UpdateComplianceTest extends SPARQL11UpdateComplianceTest {
-
-	public EndpointSPARQL11UpdateComplianceTest() {
-
-		List<String> testToIgnore = new ArrayList<>();
-		// @todo these tests are failing and should not, they are skipped so
-		// that we can be sure that we see when
-		// currently passing tests are not failing. Many of these tests are not
-		// so problematic since we do not support
-		// named graphs anyway
-		testToIgnore.add("DELETE INSERT 1b");
-		testToIgnore.add("DELETE INSERT 1c");
-		testToIgnore.add("CLEAR NAMED");
-		testToIgnore.add("DROP NAMED");
-		this.setIgnoredTests(testToIgnore);
-	}
-
-	@Rule
-	public TemporaryFolder tempDir = TemporaryFolder.builder().assureDeletion().build();
-
-	@Override
-	protected Repository newRepository() throws Exception {
-		File nativeStore = tempDir.newFolder();
-		File hdtStore = tempDir.newFolder();
-		HDTOptions spec = HDTOptions.of(HDTOptionsKeys.DICTIONARY_TYPE_KEY,
-				HDTOptionsKeys.DICTIONARY_TYPE_VALUE_MULTI_OBJECTS);
-		try (HDT hdt = Utility.createTempHdtIndex(tempDir, true, false, spec)) {
-			assert hdt != null;
-			hdt.saveToHDT(hdtStore.getAbsolutePath() + "/" + EndpointStoreTest.HDT_INDEX_NAME, null);
-		}
-
-		EndpointStore endpoint = new EndpointStore(hdtStore.getAbsolutePath() + "/", EndpointStoreTest.HDT_INDEX_NAME,
-				spec, nativeStore.getAbsolutePath() + "/", true);
-		// endpoint.setThreshold(2);
-
-		return new SailRepository(endpoint);
-		// return new DatasetRepository(new SailRepository(new
-		// NativeStore(tempDir.newFolder(), "spoc")));
-	}
-
-}
